requirements.txt
streamlit>=1.28.0
astropy>=5.3.0
astroquery>=0.4.6
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
scipy>=1.10.0
lmfit>=1.2.0
requests>=2.31.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
Pillow>=10.0.0
